Green Glass World
Write a Java program to determine whether an input word is in the "Green Glass World". Any word that is spelled with two consecutive letters of the same kind, such as "grass" or "green" or "pizza" is said to be in the world.

Your program will take one word as input from the user and return a True/False answer. Your program should only accept single words (alphabets only) and not phrases.